+++ barretenberg/cpp/src/barretenberg/vm2/simulation/events/alu_event.hpp
@@ -42,13 +42,6 @@ inline std::string to_string(AluError e)
     __builtin_unreachable();
 }
 
-class AluException : public std::runtime_error {
-  public:
-    explicit AluException(const std::string& message)
-        : std::runtime_error("ALU Exception: " + message)
-    {}
-};
-
 // Explanations on default values for b and c:
 // execution.register[X] == 0 and execution.mem_tag_reg[X] == 0 when we throw an error in execution, because
 // in the trace the default value is 0.
