+++ barretenberg/cpp/src/barretenberg/vm2/simulation/events/l1_to_l2_message_tree_check_event.hpp
@@ -2,7 +2,7 @@
 
 #include "barretenberg/vm2/common/aztec_types.hpp"
 #include "barretenberg/vm2/common/field.hpp"
-#include "barretenberg/vm2/simulation/lib/db_interfaces.hpp"
+#include "barretenberg/vm2/simulation/interfaces/db.hpp"
 
 #include <cstdint>
 #include <vector>
