+++ barretenberg/cpp/src/barretenberg/vm2/simulation/gadgets/concrete_dbs.hpp
@@ -10,7 +10,7 @@
 #include "barretenberg/vm2/simulation/gadgets/protocol_contracts.hpp"
 #include "barretenberg/vm2/simulation/gadgets/public_data_tree_check.hpp"
 #include "barretenberg/vm2/simulation/gadgets/written_public_data_slots_tree_check.hpp"
-#include "barretenberg/vm2/simulation/lib/db_interfaces.hpp"
+#include "barretenberg/vm2/simulation/interfaces/db.hpp"
 #include "barretenberg/vm2/simulation/lib/raw_data_dbs.hpp"
 
 namespace bb::avm2::simulation {
