+++ barretenberg/cpp/src/barretenberg/vm2/simulation/gadgets/public_data_tree_check.cpp
@@ -1,6 +1,6 @@
 #include "barretenberg/vm2/simulation/gadgets/public_data_tree_check.hpp"
 
-#include "barretenberg/vm2/simulation/lib/db_interfaces.hpp"
+#include "barretenberg/vm2/simulation/interfaces/db.hpp"
 
 namespace bb::avm2::simulation {
