+++ barretenberg/cpp/src/barretenberg/vm2/simulation/gadgets/tx_context.hpp
@@ -4,7 +4,7 @@
 #include "barretenberg/vm2/simulation/events/tx_context_event.hpp"
 #include "barretenberg/vm2/simulation/gadgets/context_provider.hpp"
 #include "barretenberg/vm2/simulation/gadgets/written_public_data_slots_tree_check.hpp"
-#include "barretenberg/vm2/simulation/lib/db_interfaces.hpp"
+#include "barretenberg/vm2/simulation/interfaces/db.hpp"
 
 namespace bb::avm2::simulation {
