+++ barretenberg/cpp/src/barretenberg/vm2/simulation/interfaces/nullifier_tree_check.hpp
@@ -5,6 +5,7 @@
 
 #include "barretenberg/vm2/common/aztec_types.hpp"
 #include "barretenberg/vm2/common/field.hpp"
+#include "barretenberg/vm2/simulation/lib/db_types.hpp"
 
 namespace bb::avm2::simulation {
