+++ barretenberg/cpp/src/barretenberg/vm2/simulation/lib/db_types.hpp
@@ -9,4 +9,13 @@ namespace bb::avm2::simulation {
 using NullifierTreeLeafPreimage = crypto::merkle_tree::IndexedLeaf<crypto::merkle_tree::NullifierLeafValue>;
 using PublicDataTreeLeafPreimage = crypto::merkle_tree::IndexedLeaf<crypto::merkle_tree::PublicDataLeafValue>;
 
+struct TreeCounters {
+    uint32_t note_hash_counter;
+    uint32_t nullifier_counter;
+    uint32_t l2_to_l1_msg_counter;
+    // public data tree counter is tracked via the written public data slots tree
+
+    bool operator==(const TreeCounters& other) const = default;
+};
+
 } // namespace bb::avm2::simulation
