+++ barretenberg/cpp/src/barretenberg/vm2/simulation/lib/indexed_memory_tree.hpp
@@ -3,7 +3,7 @@
 #include <cstddef>
 
 #include "barretenberg/crypto/merkle_tree/memory_tree.hpp"
-#include "barretenberg/vm2/simulation/lib/db_interfaces.hpp"
+#include "barretenberg/vm2/simulation/interfaces/db.hpp"
 
 namespace bb::avm2::simulation {
