+++ barretenberg/cpp/src/barretenberg/vm2/simulation/lib/protocol_contract_tree.hpp
@@ -4,7 +4,7 @@
 
 #include "barretenberg/vm2/common/aztec_types.hpp"
 #include "barretenberg/vm2/common/map.hpp"
-#include "barretenberg/vm2/simulation/lib/db_interfaces.hpp"
+#include "barretenberg/vm2/simulation/interfaces/db.hpp"
 #include "barretenberg/vm2/simulation/lib/indexed_memory_tree.hpp"
 
 namespace bb::avm2::simulation {
