+++ barretenberg/cpp/src/barretenberg/vm2/simulation/standalone/pure_alu.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "barretenberg/vm2/common/memory_types.hpp"
-#include "barretenberg/vm2/simulation/gadgets/alu.hpp"
+#include "barretenberg/vm2/simulation/interfaces/alu.hpp"
 
 namespace bb::avm2::simulation {
