+++ barretenberg/cpp/src/barretenberg/vm2/simulation/standalone/pure_gt.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "barretenberg/vm2/simulation/gadgets/gt.hpp"
+#include "barretenberg/vm2/simulation/interfaces/gt.hpp"
 
 #include "barretenberg/numeric/uint128/uint128.hpp"
 #include "barretenberg/numeric/uint256/uint256.hpp"
