+++ barretenberg/cpp/src/barretenberg/vm2/simulation/standalone/pure_to_radix.cpp
@@ -7,6 +7,7 @@
 #include "barretenberg/common/bb_bench.hpp"
 #include "barretenberg/numeric/uint256/uint256.hpp"
 #include "barretenberg/vm2/common/to_radix.hpp"
+#include "barretenberg/vm2/simulation/interfaces/memory.hpp"
 
 namespace bb::avm2::simulation {
